decomp: refactor and cleanup unimplementation macro
